+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed parts of the game update function to handle new files
+- Fixed a panic when the launcher folder is a broken symlink, the launcher will exit with an
+  error message instead
 
 ## [1.14.4] - 2026-02-13