+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for layered and animated backgrounds. You can disable video background in settings.
+
 ### Fixed
 
 - Fixed parts of the game update function to handle new files